Suggest using chrony as NTP daemon
RedHat strongly suggests [0] using chrony instead of ntpd and switched the
default:

17.1.2. Choosing Between NTP Daemons
Chrony should be preferred for all systems except for the systems that are
managed or monitored by tools that do not support chrony, or the systems that
have a hardware reference clock which cannot be used with chrony.

Feature comparison: [1] and [2].

It's available on RedHat/CentOS as well as Debian/Ubuntu distributions.

[0] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/system_administrators_guide/ch-configuring_ntp_using_the_chrony_suite
[1] https://chrony.tuxfamily.org/comparison.html
[2] https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/7/html/system_administrators_guide/ch-configuring_ntp_using_the_chrony_suite#sect-differences_between_ntpd_and_chronyd