Require configuring the NIC of VM as promiscuous
This change applies to working with a nested virtual environment. 

CloudStack's system VMs rely on the deployment VM for Internet access.
When the NIC of the deployment VM is not promiscuous, it will deny traffic from the system VMs, 
and thus, the system VMs cannot reach the Internet. This will create a lot of issues after CloudStack is installed.
It's better to explicitly mention this in the installation guide, so that people will use the correct 
setting from the beginning, instead of spending a lot of time to realize this is the problem.